fix(generator): move utils to utils module
