+++ pixels/generator/stac_utils.py
@@ -6,10 +6,14 @@ import shutil
 from urllib.parse import urlparse
 
 import boto3
+import numpy as np
 import pystac
+import rasterio
 import sentry_sdk
 import structlog
 from pystac import STAC_IO
+
+from pixels.exceptions import PixelsException
 
 logger = structlog.get_logger(__name__)
 
@@ -129,3 +133,124 @@ def open_file_from_s3(source_path):
         logger.warning(f"s3.exceptions.NoSuchKey. source_path {source_path}")
         data = None
     return data
+
+
+def list_files_in_folder(uri, filetype="tif"):
+    parsed = urlparse(uri)
+    if parsed.scheme == "s3":
+        return list_files_in_s3(uri, filetype=filetype)
+    else:
+        return glob.glob(f"{uri}/**{filetype}", recursive=True)
+
+
+def check_file_in_s3(uri):
+    """
+    Check if file exists at an S3 uri.
+
+    Parameters
+    ----------
+    uri: str
+        The S3 uri to check if file exists. Example: s3://my-bucket/config.json
+    """
+    # Split the S3 uri into compoments.
+    parsed = urlparse(uri)
+    # Ensure input is a s3 uri.
+    if parsed.scheme != "s3":
+        raise PixelsException("Invalid S3 uri found: {}.".format(uri))
+    # Get bucket name.
+    bucket = parsed.netloc
+    # Get key in bucket.
+    key = parsed.path[1:]
+    # List objects with that key.
+    s3 = boto3.client("s3")
+    theObjs = s3.list_objects_v2(Bucket=bucket, Prefix=os.path.dirname(key))
+    list_obj = [ob["Key"] for ob in theObjs["Contents"]]
+    # Ensure key is in list.
+    return key in list_obj
+
+
+def upload_obj_s3(uri, obj):
+    parsed = urlparse(uri)
+    if parsed.scheme == "s3":
+        bucket = parsed.netloc
+        key = parsed.path[1:]
+        s3 = boto3.client("s3")
+        s3.put_object(Key=key, Bucket=bucket, Body=obj)
+
+
+def list_files_in_s3(uri, filetype="tif"):
+    parsed = urlparse(uri)
+    if parsed.scheme == "s3":
+        bucket = parsed.netloc
+        key = parsed.path[1:]
+        s3 = boto3.client("s3")
+        paginator = s3.get_paginator("list_objects_v2")
+        theObjs = paginator.paginate(Bucket=bucket, Prefix=key)
+        # Get list of objects if thre are any.
+        mult_obj = [ob["Contents"] for ob in theObjs if "Contents" in ob]
+        list_obj = []
+        for obj in mult_obj:
+            ob = [
+                "s3://" + bucket + "/" + f["Key"]
+                for f in obj
+                if f["Key"].endswith(filetype)
+            ]
+            list_obj = list_obj + ob
+    return list_obj
+
+
+def get_bbox_and_footprint_and_stats(raster_uri, categorical):
+    """with open(path, "r") as file:
+        file.write(json.dumps(catalog_dict))
+    Get bounding box and footprint from raster.
+
+    Parameters
+    ----------
+    raster_uri : str or bytes_io
+        The raster file location or bytes_io.
+    categorical: boolean, optional
+        If True, compute statistics of the pixel data for class weighting.
+
+    Returns
+    -------
+    bbox : list
+        Bounding box of input raster.
+    footprint : list
+        Footprint of input raster.
+    datetime_var : datetime type
+        Datetime from image.
+    out_meta : rasterio meta type
+        Metadata from raster.
+    stats: dict or None
+        Statistics of the data, counts by unique value.
+    """
+    with rasterio.open(raster_uri) as ds:
+        # Get bounds.
+        bounds = ds.bounds
+        # Create bbox as list.
+        bbox = [bounds.left, bounds.bottom, bounds.right, bounds.top]
+        # Create bbox as polygon feature.
+        footprint = {
+            "type": "Polygon",
+            "coordinates": [
+                [
+                    [bounds.left, bounds.bottom],
+                    [bounds.left, bounds.top],
+                    [bounds.right, bounds.top],
+                    [bounds.right, bounds.bottom],
+                    [bounds.left, bounds.bottom],
+                ]
+            ],
+        }
+        # Try getting the datetime in the raster metadata. Set to None if not
+        # found.
+        datetime_var = ds.tags().get("datetime", None)
+        # Compute unique counts if requested.
+        stats = None
+        if categorical:
+            unique_values, uniue_counts = np.unique(ds.read(), return_counts=True)
+            stats = {
+                int(key): int(val) for key, val in zip(unique_values, uniue_counts)
+            }
+
+        return bbox, footprint, datetime_var, ds.meta, stats
